Create e2e/test_play.py
